refactor(frontend): make useSessionId context-backed for deep access
