+++ apps/frontend/src/pages/chat/ChatPage.tsx
@@ -5,6 +5,7 @@ import {useAutoScroll} from '@/hooks/useAutoScroll.js';
 import {ChatPageView} from './ChatPageView.js';
 import {ChatEventBusProvider} from './contexts/ChatEventBusContext/index.js';
 import {SessionConfigProvider} from './contexts/SessionConfigContext/index.js';
+import {SessionIdProvider} from './contexts/SessionIdContext/index.js';
 import {
   ToolOutputProvider,
   useClearToolOutput,
@@ -19,13 +20,15 @@ import {useStreamChat} from './hooks/useStreamChat.js';
 /** Chat page container. Wraps content in providers. */
 export function ChatPage() {
   return (
-    <ChatEventBusProvider>
-      <SessionConfigProvider>
-        <ToolOutputProvider>
-          <ChatPageContent />
-        </ToolOutputProvider>
-      </SessionConfigProvider>
-    </ChatEventBusProvider>
+    <SessionIdProvider>
+      <ChatEventBusProvider>
+        <SessionConfigProvider>
+          <ToolOutputProvider>
+            <ChatPageContent />
+          </ToolOutputProvider>
+        </SessionConfigProvider>
+      </ChatEventBusProvider>
+    </SessionIdProvider>
   );
 }
