+++ apps/frontend/src/pages/chat/contexts/SessionIdContext/index.ts
@@ -1,0 +1,5 @@
+export {
+  SessionIdContext,
+  type SessionIdContextValue,
+} from './SessionIdContext.js';
+export {SessionIdProvider} from './SessionIdProvider.js';
