+++ apps/frontend/src/pages/chat/hooks/useSessionId.ts
@@ -1,46 +1,18 @@
-import {useCallback, useState} from 'react';
+import {use} from 'react';
 
-import {createSession} from '@/api/chat/index.js';
+import {
+  SessionIdContext,
+  type SessionIdContextValue,
+} from '../contexts/SessionIdContext/index.js';
 
-interface SessionConfig {
-  workspace?: string;
-  extraAllowedPaths?: readonly string[];
-}
-
-/** Manages the session ID lifecycle. Session is created on demand via `createNewSessionId`. */
-export function useSessionId() {
-  const [sessionId, setSessionId] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-
-  const createNewSessionId = useCallback(async (config: SessionConfig = {}) => {
-    setError(null);
-    try {
-      const id = await createSession(config);
-      setSessionId(id);
-      return id;
-    } catch (e) {
-      console.error('Failed to create session', e);
-      const message =
-        e instanceof Error ? e.message : 'Failed to create session';
-      setError(message);
-      return null;
-    }
-  }, []);
-
-  const clearSessionId = useCallback(() => {
-    setSessionId(null);
-    setError(null);
-  }, []);
-
-  const clearCreateNewSessionIdError = useCallback(() => {
-    setError(null);
-  }, []);
-
-  return {
-    sessionId,
-    createNewSessionIdError: error,
-    createNewSessionId,
-    clearSessionId,
-    clearCreateNewSessionIdError,
-  };
+/**
+ * Returns the session ID and lifecycle methods.
+ * Must be used within a SessionIdProvider.
+ */
+export function useSessionId(): SessionIdContextValue {
+  const value = use(SessionIdContext);
+  if (!value) {
+    throw new Error('useSessionId must be used within a SessionIdProvider');
+  }
+  return value;
 }
